add location dependecies and update colors

+++ app/src/main/java/com/juanfe/project/weatherapp/data/network/response/Responses.kt
@@ -25,13 +25,13 @@ data class LocationResponse(
     @SerializedName("tz_id")
     val tzId: String,
     @SerializedName("localtime_epoch")
-    val localtimeEpoch: Long,
+    val localtimeEpoch: Double,
     val localtime: String,
 )
 
 data class CurrentResponse(
     @SerializedName("last_updated_epoch")
-    val lastUpdatedEpoch: Long,
+    val lastUpdatedEpoch: Double,
     @SerializedName("last_updated")
     val lastUpdated: String,
     @SerializedName("temp_c")
@@ -39,26 +39,26 @@ data class CurrentResponse(
     @SerializedName("temp_f")
     val tempF: Double,
     @SerializedName("is_day")
-    val isDay: Long,
+    val isDay: Double,
     val condition: ConditionResponse,
     @SerializedName("wind_mph")
     val windMph: Double,
     @SerializedName("wind_kph")
     val windKph: Double,
     @SerializedName("wind_degree")
-    val windDegree: Long,
+    val windDegree: Double,
     @SerializedName("wind_dir")
     val windDir: String,
     @SerializedName("pressure_mb")
-    val pressureMb: Long,
+    val pressureMb: Double,
     @SerializedName("pressure_in")
     val pressureIn: Double,
     @SerializedName("precip_mm")
     val precipMm: Double,
     @SerializedName("precip_in")
     val precipIn: Double,
-    val humidity: Long,
-    val cloud: Long,
+    val humidity: Double,
+    val cloud: Double,
     @SerializedName("feelslike_c")
     val feelsLikeC: Double,
     @SerializedName("feelslike_f")
@@ -76,9 +76,9 @@ data class CurrentResponse(
     @SerializedName("dewpoint_f")
     val dewPointF: Double,
     @SerializedName("vis_km")
-    val visKm: Long,
+    val visKm: Double,
     @SerializedName("vis_miles")
-    val visMiles: Long,
+    val visMiles: Double,
     val uv: Double,
     @SerializedName("gust_mph")
     val gustMph: Double,
@@ -89,7 +89,7 @@ data class CurrentResponse(
 data class ConditionResponse(
     val text: String,
     val icon: String,
-    val code: Long,
+    val code: Double,
 )
 
 data class ForecastResponse(
@@ -100,7 +100,7 @@ data class ForecastResponse(
 data class ForecastDayResponse(
     val date: String,
     @SerializedName("date_epoch")
-    val dateEpoch: Long,
+    val dateEpoch: Double,
     val day: DayResponse,
     val astro: AstroResponse,
     val hour: List<HourResponse>,
@@ -128,20 +128,20 @@ data class DayResponse(
     @SerializedName("totalprecip_in")
     val totalPrecipIn: Double,
     @SerializedName("totalsnow_cm")
-    val totalSnowCm: Long,
+    val totalSnowCm: Double,
     @SerializedName("avgvis_km")
     val avgVisKm: Double,
     @SerializedName("avgvis_miles")
-    val avgVisMiles: Long,
-    val avgHumidity: Long,
+    val avgVisMiles: Double,
+    val avgHumidity: Double,
     @SerializedName("daily_will_it_rain")
-    val dailyWillItRain: Long,
+    val dailyWillItRain: Double,
     @SerializedName("daily_chance_of_rain")
-    val dailyChanceOfRain: Long,
+    val dailyChanceOfRain: Double,
     @SerializedName("daily_will_it_snow")
-    val dailyWillItSnow: Long,
+    val dailyWillItSnow: Double,
     @SerializedName("daily_chance_of_snow")
-    val dailyChanceOfSnow: Long,
+    val dailyChanceOfSnow: Double,
     val condition: Condition2Response,
     val uv: Double,
 )
@@ -149,7 +149,7 @@ data class DayResponse(
 data class Condition2Response(
     val text: String,
     val icon: String,
-    val code: Long,
+    val code: Double,
 )
 
 data class AstroResponse(
@@ -161,34 +161,34 @@ data class AstroResponse(
     @SerializedName("moon_phase")
     val moonPhase: String,
     @SerializedName("moon_illumination")
-    val moonIllumination: Long,
+    val moonIllumination: Double,
     @SerializedName("is_moon_up")
-    val isMoonUp: Long,
+    val isMoonUp: Double,
     @SerializedName("is_sun_up")
-    val isSunUp: Long,
+    val isSunUp: Double,
 )
 
 data class HourResponse(
     @SerializedName("time_epoch")
-    val timeEpoch: Long,
+    val timeEpoch: Double,
     val time: String,
     @SerializedName("temp_c")
     val tempC: Double,
     @SerializedName("temp_f")
     val tempF: Double,
     @SerializedName("is_day")
-    val isDay: Long,
+    val isDay: Double,
     val condition: Condition3Response,
     @SerializedName("wind_mph")
     val windMph: Double,
     @SerializedName("wind_kph")
     val windKph: Double,
     @SerializedName("wind_degree")
-    val windDegree: Long,
+    val windDegree: Double,
     @SerializedName("wind_dir")
     val windDir: String,
     @SerializedName("pressure_mb")
-    val pressureMb: Long,
+    val pressureMb: Double,
     @SerializedName("pressure_in")
     val pressureIn: Double,
     @SerializedName("precip_mm")
@@ -196,9 +196,9 @@ data class HourResponse(
     @SerializedName("precip_in")
     val precipIn: Double,
     @SerializedName("snow_cm")
-    val snowCm: Long,
-    val humidity: Long,
-    val cloud: Long,
+    val snowCm: Double,
+    val humidity: Double,
+    val cloud: Double,
     @SerializedName("feelslike_c")
     val feelsLikeC: Double,
     @SerializedName("feelslike_f")
@@ -216,17 +216,17 @@ data class HourResponse(
     @SerializedName("dewpoint_f")
     val dewPointF: Double,
     @SerializedName("will_it_rain")
-    val willItRain: Long,
+    val willItRain: Double,
     @SerializedName("chance_of_rain")
-    val chanceOfRain: Long,
+    val chanceOfRain: Double,
     @SerializedName("will_it_snow")
-    val willItSnow: Long,
+    val willItSnow: Double,
     @SerializedName("chance_of_snow")
-    val chanceOfSnow: Long,
+    val chanceOfSnow: Double,
     @SerializedName("vis_km")
     val visKm: Double,
     @SerializedName("vis_miles")
-    val visMiles: Long,
+    val visMiles: Double,
     @SerializedName("gust_mph")
     val gustMph: Double,
     @SerializedName("gust_kph")
@@ -237,5 +237,5 @@ data class HourResponse(
 data class Condition3Response(
     val text: String,
     val icon: String,
-    val code: Long,
+    val code: Double,
 )
